optimized performance with native buffers
